Create files for the widgets replacing CardScreen
