+++ lib/music_player/music_player_screen.dart
@@ -1,11 +1,14 @@
 import 'package:flutter/material.dart';
-import 'package:musbx/card_screen.dart';
 import 'package:musbx/music_player/loop_card/loop_card.dart';
 import 'package:musbx/music_player/music_player.dart';
 import 'package:musbx/music_player/pitch_speed_card/pitch_speed_card.dart';
 import 'package:musbx/music_player/position_card/button_panel.dart';
 import 'package:musbx/music_player/current_song_card/current_song_panel.dart';
 import 'package:musbx/music_player/position_card/position_slider.dart';
+import 'package:musbx/screen/card_list.dart';
+import 'package:musbx/screen/default_app_bar.dart';
+import 'package:musbx/screen/empty_tab_bar.dart';
+import 'package:musbx/screen/widget_card.dart';
 
 class MusicPlayerScreen extends StatelessWidget {
   /// Screen that allows the user to select and play a song.
